fix(frontend):  Websocket connections issue
